+++ src/main/java/dao/entity/RegisterSales.java
@@ -1,0 +1,17 @@
+package dao.entity;
+
+import java.sql.Date;
+
+public class RegisterSales {
+	String orderID;
+	String registerID;
+	String cashier;
+	Date transactionDate;
+	
+	public RegisterSales(String orderID, String registerID, String cashier, Date transactionDate) {
+		this.orderID = orderID;
+		this.registerID = registerID;
+		this.cashier = cashier;
+		this.transactionDate = transactionDate;
+	}
+}