+++ src/main/java/register/Register.java
@@ -10,6 +10,8 @@ import java.text.DateFormat;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 
+import dao.entity.Database;
+
 
 public class Register {